Move the IS_TESTING method out of settings
